feat(explorer): Epoch page. Validators. Merge with another PR.
