+++ apps/explorer/src/lib/ui/utils/index.ts
@@ -3,4 +3,5 @@
 
 export * from './generateTableDataFromCheckpointsData';
 export * from './generateTableDataFromEpochsData';
+export * from './generateValidatorsTableData';
 export * from './objectField';
