+++ apps/explorer/src/pages/validators/Validators.tsx
@@ -2,289 +2,25 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
+import React, { type JSX, useMemo } from 'react';
+import { roundFloat, useFormatCoin, useGetValidatorsApy, useGetValidatorsEvents } from '@iota/core';
 import {
     DisplayStats,
     DisplayStatsSize,
     DisplayStatsType,
     TooltipPosition,
 } from '@iota/apps-ui-kit';
-import {
-    type ApyByValidator,
-    formatAmount,
-    formatPercentageDisplay,
-    roundFloat,
-    useFormatCoin,
-    useGetValidatorsApy,
-    useGetValidatorsEvents,
-} from '@iota/core';
 import { useIotaClientQuery } from '@iota/dapp-kit';
-import { type IotaEvent, type IotaValidatorSummary } from '@iota/iota-sdk/client';
-import { Text } from '@iota/ui';
-import React, { type JSX, useMemo } from 'react';
-
-import { ErrorBoundary, PageLayout, StakeColumn } from '~/components';
 import { IOTA_TYPE_ARG } from '@iota/iota-sdk/utils';
-import { Banner, Link, PlaceholderTable, TableCard, TableHeader, Tooltip } from '~/components/ui';
-import { ampli, getValidatorMoveEvent } from '~/lib/utils';
-import { type ColumnDef } from '@tanstack/react-table';
-import { Badge, BadgeType, type TableCellProps, TableCellType } from '@iota/apps-ui-kit';
-import { VALIDATOR_LOW_STAKE_GRACE_PERIOD } from '~/lib';
-
-export interface ValidatorTableRow {
-    name: TableCellProps;
-    stake: TableCellProps;
-    apy:
-        | TableCellProps
-        | {
-              apy: number | null;
-              isApyApproxZero?: boolean;
-          };
-    nextEpochGasPrice: TableCellProps;
-    commission: TableCellProps;
-    img: string | null;
-    address: string;
-    lastReward:
-        | TableCellProps
-        | {
-              lastReward: number | null;
-          };
-    votingPower: TableCellProps;
-    atRisk: TableCellProps;
-    symbol?: string;
-}
-
-interface ValidatorsTableDataArgs {
-    validators: IotaValidatorSummary[];
-    atRiskValidators: [string, string][];
-    validatorEvents: IotaEvent[];
-    rollingAverageApys: ApyByValidator | null;
-    columns?: ColumnDef<object, unknown>[];
-    symbol?: string;
-}
-
-function generateValidatorName(address: string, name: string, imageUrl: string) {
-    return (
-        <Link
-            to={`/validator/${encodeURIComponent(address)}`}
-            onClick={() =>
-                ampli.clickedValidatorRow({
-                    sourceFlow: 'Epoch details',
-                    validatorAddress: address,
-                    validatorName: name,
-                })
-            }
-        >
-            <div className="flex items-center gap-x-2.5 text-neutral-40 dark:text-neutral-60">
-                <span className="text-label-lg">{name}</span>
-            </div>
-        </Link>
-    );
-}
-
-function generateAtRiskValidator(atRisk: number | null) {
-    const label = 'At Risk';
-    return typeof atRisk === 'number' ? (
-        <Tooltip
-            tip="Staked IOTA is below the minimum IOTA stake threshold to remain a validator."
-            onOpen={() =>
-                ampli.activatedTooltip({
-                    tooltipLabel: label,
-                })
-            }
-        >
-            <div className="flex cursor-pointer flex-nowrap items-center">
-                <div className="text-body-md uppercase text-neutral-40">{label}</div>
-                &nbsp;
-                <div className="text-body-md uppercase text-neutral-40">
-                    {atRisk > 1 ? `in ${atRisk} epochs` : 'next epoch'}
-                </div>
-            </div>
-        </Tooltip>
-    ) : (
-        <Badge type={BadgeType.Neutral} label="Active" />
-    );
-}
-
-export function generateValidatorsTableData({
-    validators,
-    atRiskValidators,
-    validatorEvents,
-    rollingAverageApys,
-    columns,
-    symbol,
-}: ValidatorsTableDataArgs): { data: ValidatorTableRow[]; columns: ColumnDef<object, unknown>[] } {
-    return {
-        data: [...validators]
-            .sort(() => 0.5 - Math.random())
-            .map((validator) => {
-                const validatorName = validator.name;
-                const totalStake = validator.stakingPoolIotaBalance;
-                const img = validator.imageUrl;
-
-                const event = getValidatorMoveEvent(validatorEvents, validator.iotaAddress) as {
-                    pool_staking_reward?: string;
-                };
-
-                const atRiskValidator = atRiskValidators.find(
-                    ([address]) => address === validator.iotaAddress,
-                );
-                const isAtRisk = !!atRiskValidator;
-                const lastReward = event?.pool_staking_reward ?? null;
-                const { apy, isApyApproxZero } = rollingAverageApys?.[validator.iotaAddress] ?? {
-                    apy: null,
-                };
-
-                return {
-                    name: {
-                        type: TableCellType.Children,
-                        children: generateValidatorName(
-                            validator.iotaAddress,
-                            validatorName,
-                            validator.imageUrl,
-                        ),
-                    },
-                    stake: {
-                        type: TableCellType.Text,
-                        label: formatAmount(totalStake),
-                        supportingLabel: symbol,
-                        stake: totalStake,
-                    },
-                    apy: {
-                        type: TableCellType.AvatarText,
-                        label: formatPercentageDisplay(apy, '--', isApyApproxZero),
-                        apy,
-                        isApyApproxZero,
-                    },
-                    nextEpochGasPrice: {
-                        type: TableCellType.Text,
-                        label: formatAmount(validator.nextEpochGasPrice),
-                        supportingLabel: symbol,
-                        nextEpochGasPrice: validator.nextEpochGasPrice,
-                    },
-                    commission: {
-                        type: TableCellType.Text,
-                        label: `${Number(validator.commissionRate) / 100}%`,
-                    },
-                    img: img,
-                    address: validator.iotaAddress,
-                    lastReward: {
-                        type: TableCellType.Text,
-                        label: lastReward ? formatAmount(Number(lastReward)) : '--',
-                        supportingLabel: lastReward ? symbol : undefined,
-                        lastReward: lastReward ?? null,
-                    },
-                    votingPower: {
-                        type: TableCellType.Text,
-                        label: validator.votingPower
-                            ? Number(validator.votingPower) / 100 + '%'
-                            : '--',
-                    },
-                    atRisk: {
-                        type: TableCellType.Children,
-                        children: generateAtRiskValidator(
-                            isAtRisk
-                                ? VALIDATOR_LOW_STAKE_GRACE_PERIOD - Number(atRiskValidator[1])
-                                : null,
-                        ),
-                    },
-                    symbol: symbol,
-                };
-            }),
-        columns: columns || [
-            {
-                header: '#',
-                accessorKey: 'number',
-                meta: {
-                    isRenderCell: true,
-                },
-                cell: (props: any) => (
-                    <div className="text-body-md text-neutral-40">
-                        {props.table.getSortedRowModel().flatRows.indexOf(props.row) + 1}
-                    </div>
-                ),
-            },
-            {
-                header: 'Name',
-                accessorKey: 'name',
-                enableSorting: true,
-                sortingFn: (a: any, b: any, colId: string) =>
-                    a.getValue(colId).name.localeCompare(b.getValue(colId).name, 'en', {
-                        sensitivity: 'base',
-                        numeric: true,
-                    }),
-            },
-            {
-                header: 'Stake',
-                accessorKey: 'stake',
-                enableSorting: true,
-                meta: {
-                    isRenderCell: true,
-                },
-                cell: (props: any) => <StakeColumn stake={props.getValue().stake} />,
-            },
-            {
-                header: 'Proposed Next Epoch Gas Price',
-                accessorKey: 'nextEpochGasPrice',
-                enableSorting: true,
-                meta: {
-                    isRenderCell: true,
-                },
-                cell: (props: any) => (
-                    <StakeColumn stake={props.getValue().nextEpochGasPrice} inNano />
-                ),
-            },
-            {
-                header: 'APY',
-                accessorKey: 'apy',
-                enableSorting: true,
-                meta: {
-                    isRenderCell: true,
-                },
-                sortingFn: (a: any, b: any, colId: string) =>
-                    a.getValue(colId)?.apy < b.getValue(colId)?.apy ? -1 : 1,
-                cell: (props: any) => {
-                    const { apy, isApyApproxZero } = props.getValue();
-                    return (
-                        <div className="text-body-md text-neutral-40">
-                            {formatPercentageDisplay(apy, '--', isApyApproxZero)}
-                        </div>
-                    );
-                },
-            },
-            {
-                header: 'Commission',
-                accessorKey: 'commission',
-                enableSorting: true,
-            },
-            {
-                header: 'Last Epoch Rewards',
-                accessorKey: 'lastReward',
-                enableSorting: true,
-                cell: (props: any) => {
-                    const lastReward = props.getValue().lastReward;
-                    return lastReward !== null ? (
-                        <StakeColumn stake={Number(lastReward)} />
-                    ) : (
-                        <Text variant="bodySmall/medium" color="steel-darker">
-                            --
-                        </Text>
-                    );
-                },
-            },
-            {
-                header: 'Voting Power',
-                accessorKey: 'votingPower',
-                enableSorting: true,
-            },
-            {
-                header: 'Status',
-                accessorKey: 'atRisk',
-            },
-        ],
-    };
-}
+import {
+    ErrorBoundary,
+    PageLayout,
+    Banner,
+    PlaceholderTable,
+    TableCard,
+    TableHeader,
+} from '~/components';
+import { generateValidatorsTableData } from '~/lib/ui/utils/generateValidatorsTableData';
 
 function ValidatorPageResult(): JSX.Element {
     const { data, isPending, isSuccess, isError } = useIotaClientQuery('getLatestIotaSystemState');
